requirements.txt
transformers
transformers[sentencepiece]
sentence-transformers
datasets
beautifulsoup4
scikit-learn
sacrebleu
scipy
rouge_score 
py7zr
pandas
nltk
tqdm
PyYAML
matplotlib
notebook
boto3
mypy-boto3-s3
python-box==6.0.2
ensure==1.0.2
flask
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.2
-e .